feat(feature:loan): migrated to CMP (#2786)

+++ core/model/src/commonMain/kotlin/org/mifos/mobile/core/model/entity/accounts/loan/LoanWithAssociations.kt
@@ -61,7 +61,7 @@ data class LoanWithAssociations(
 
     val repaymentFrequencyType: RepaymentFrequencyType? = null,
 
-    val interestRatePerPeriod: Int? = null,
+    val interestRatePerPeriod: Double? = null,
 
     val interestRateFrequencyType: InterestRateFrequencyType? = null,
